+++ common-include/include/mswin/vaDbgTs.CHHI.h
@@ -20,7 +20,7 @@ inline Uint64 get_qpf()
 	LARGE_INTEGER li = {};
 	BOOL succ = QueryPerformanceFrequency(&li);
 	if(!succ)
-		return -1;
+		return (Uint64)(__int64)-1;
 	else
 		return li.QuadPart;
 }
@@ -30,7 +30,7 @@ inline Uint64 get_qpc()
 	LARGE_INTEGER li = {};
 	BOOL succ = QueryPerformanceCounter(&li);
 	if(!succ)
-		return -1;
+		return (Uint64)(__int64)-1;
 	else
 		return li.QuadPart;
 }
